Temporarily disable checking for default types
It looks like, currently, types aren't being followed through to their
defines and we assume that objects in general don't have a type
(undefined). Because of this, if the return type is !defined and then
referenced as a type later for an object then we won't be able to see
that object because it will be seen as a non-default type. This
temporarily works around that by just disabling this check entirely.